pyproject.toml
[tool.poetry]
name = "flask-demo"
version = "0.1.0"
description = ""
authors = ["liujianwei1 <liujianwei1@tal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-wtf = "^1.2.1"
cymysql = "^1.0.1"
flask-cors = "^4.0.1"
requests = "^2.32.3"
flask-sqlalchemy = "^3.1.1"
email-validator = "^2.1.1"
flask-httpauth = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
